virtual-nascom: switch to SDL2 and stop it crashing on dwim keyboard
The video is done so that a bitmapped display is trivial to merge with the
current set up. That will be important for the various graphics cards that
could be added to the Nascom.

The crashes were caused by passing very large values to isalpha and toupper.
They take int inputs so are not type clipped but are only valid for 00-FF and
EOF.