Fix #11 - attempt to match speed of real Nascom and bump to 60 Hz
Yaze, the Z80 simulator we are using isn't accounting for T states.
Until we fix that, assume instructions are 8 T-states on average and
use that to improve the faithfulness of the simulation (it feels about
right).  Also, bump the UI speed to 60 Hz as recommended by SDL (Thanks
to Peter Jensen for that suggestion).